Move coprocessor test utilites to test_coprocessor library (#3495)
