+++ components/test_coprocessor/lib.rs
@@ -1,0 +1,35 @@
+// Copyright 2018 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+extern crate kvproto;
+extern crate tipb;
+
+extern crate futures;
+extern crate protobuf;
+
+extern crate test_storage;
+extern crate tikv;
+
+mod column;
+mod dag;
+mod fixture;
+mod store;
+mod table;
+mod util;
+
+pub use column::*;
+pub use dag::*;
+pub use fixture::*;
+pub use store::*;
+pub use table::*;
+pub use util::*;
